+++ mirrord-config/src/lib.rs
@@ -256,6 +256,7 @@ mod tests {
                 ttl: Some(60),
                 ephemeral: Some(false),
                 communication_timeout: None,
+                startup_timeout: None,
             },
             feature: FeatureFileConfig {
                 env: ToggleableConfig::Enabled(true),