+++ CHANGELOG.md
@@ -6,10 +6,14 @@ Previous versions had CHANGELOG per component, we decided to combine all reposit
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
+### Added
+- Telemetries, see [TELEMETRY.md](./TELEMETRY.md) for more information.
 
 ### Added
 
 - Add changelog for intelliJ extension, closes [[#542](https://github.com/metalbear-co/mirrord/issues/542)]
+- - Added timeout for "waiting for pod to be ready..." in mirrord-layer to prevent unresponsive behavior. See [#579](https://github.com/metalbear-co/mirrord/issues/579)
+- IntelliJ Extension: Default log level to `ERROR` from `DEBUG`
 
 ### Changed